CODEVIZ-2.0/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
google-genai>=0.1.0
chromadb>=0.4.0
email-validator>=2.0.0
